requirements.txt
fastapi
uvicorn[standard]
starlette
langchain
langchain-community
langchain-postgres
psycopg[binary] 
pandas
python-dotenv
python-multipart
requests
sqlalchemy
loguru
openai
tiktoken
ollama
langchain-ollama
